Replace direct drawing mode from SDLGLOutputSurface
Similar to many of the previous patches. This code is used by the SDLGL_FB16
and SDLGL_FB32 renderers (those renderers may disappear in the future).

+++ src/video/SDLGLOutputSurface.hh
@@ -35,7 +35,7 @@ protected:
 	                    unsigned width, unsigned height);
 
 private:
-	double texCoordX, texCoordY;
+	float texCoordX, texCoordY;
 	gl::Texture fbTex;
 	MemBuffer<char> fbBuf;
 	const FrameBuffer frameBuffer;
